Update for smaller objects

+++ SlidingPuzzles/src/AStarPriority.java
@@ -1,6 +1,4 @@
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.PriorityQueue;
 
 
@@ -21,10 +19,10 @@ public class AStarPriority {
 			System.out.println("Sorry, it looks like this puzzle is IMPOSSIBLE to solve.");
 			return null;
 		}
-		int row = 0;
-		int col = 0;
-		for (int i=0; i<b.getBoard().length; i++) {
-			for (int j=0; j<b.getBoard().length; j++) {
+		byte row = 0;
+		byte col = 0;
+		for (byte i=0; i<b.getBoard().length; i++) {
+			for (byte j=0; j<b.getBoard().length; j++) {
 				if (b.getBoard()[i][j] == 0) {
 					row = i;
 					col = j;
